Extract betaDialog and courseList
