+++ feature/dashboard/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/dashboard/ui/CoursesOverview.kt
@@ -1,15 +1,11 @@
 package de.tum.informatics.www1.artemis.native_app.feature.dashboard.ui
 
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -20,17 +16,12 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberTopAppBarState
@@ -39,38 +30,22 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.min
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.compose.composable
-import de.tum.informatics.www1.artemis.native_app.core.model.Course
-import de.tum.informatics.www1.artemis.native_app.core.model.CourseWithScore
 import de.tum.informatics.www1.artemis.native_app.core.model.Dashboard
 import de.tum.informatics.www1.artemis.native_app.core.ui.common.BasicDataStateUi
-import de.tum.informatics.www1.artemis.native_app.core.ui.common.course.CompactCourseHeaderViewMode
-import de.tum.informatics.www1.artemis.native_app.core.ui.common.course.CompactCourseItemHeader
-import de.tum.informatics.www1.artemis.native_app.core.ui.common.course.CourseExerciseAndLectureCount
-import de.tum.informatics.www1.artemis.native_app.core.ui.common.course.CourseItemGrid
-import de.tum.informatics.www1.artemis.native_app.core.ui.common.course.ExpandedCourseItemHeader
-import de.tum.informatics.www1.artemis.native_app.core.ui.exercise.CoursePointsDecimalFormat
 import de.tum.informatics.www1.artemis.native_app.feature.dashboard.BuildConfig
 import de.tum.informatics.www1.artemis.native_app.feature.dashboard.R
 import de.tum.informatics.www1.artemis.native_app.feature.dashboard.service.BetaHintService
@@ -78,7 +53,6 @@ import de.tum.informatics.www1.artemis.native_app.feature.dashboard.service.Surv
 import kotlinx.coroutines.launch
 import org.koin.androidx.compose.getViewModel
 import org.koin.compose.koinInject
-import java.text.DecimalFormat
 
 const val DASHBOARD_DESTINATION = "dashboard"
 internal const val TEST_TAG_COURSE_LIST = "TEST_TAG_COURSE_LIST"
@@ -247,266 +221,7 @@ internal fun CoursesOverview(
     }
 }
 
-@Composable
-private fun BetaHintDialog(
-    dismiss: (dismissPermanently: Boolean) -> Unit
-) {
-    var isDismissPersistentlyChecked by remember { mutableStateOf(false) }
 
-    AlertDialog(
-        onDismissRequest = { dismiss(false) },
-        title = { Text(text = stringResource(id = R.string.dashboard_dialog_beta_title)) },
-        text = {
-            Column(
-                modifier = Modifier,
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Text(text = stringResource(id = R.string.dashboard_dialog_beta_message))
-
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable(
-                            role = Role.Checkbox,
-                            onClick = { isDismissPersistentlyChecked = !isDismissPersistentlyChecked }
-                        ),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Checkbox(
-                        modifier = Modifier,
-                        checked = isDismissPersistentlyChecked,
-                        onCheckedChange = { isDismissPersistentlyChecked = it }
-                    )
-
-                    Text(text = stringResource(id = R.string.dashboard_dialog_beta_do_not_show_again))
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(
-                onClick = { dismiss(isDismissPersistentlyChecked) }
-            ) {
-                Text(text = stringResource(id = R.string.dashboard_dialog_beta_positive))
-            }
-        }
-    )
-}
-
-/**
- * Displays a lazy list of all the courses supplied.
- */
-@Composable
-private fun CourseList(
-    modifier: Modifier,
-    courses: List<CourseWithScore>,
-    serverUrl: String,
-    authorizationToken: String,
-    onClickOnCourse: (Course) -> Unit
-) {
-    CourseItemGrid(
-        modifier = modifier,
-        courses = courses,
-    ) { dashboardCourse, courseItemModifier, isCompact ->
-        CourseItem(
-            modifier = courseItemModifier.testTag(testTagForCourse(dashboardCourse.course.id!!)),
-            courseWithScore = dashboardCourse,
-            serverUrl = serverUrl,
-            authorizationToken = authorizationToken,
-            onClick = { onClickOnCourse(dashboardCourse.course) },
-            isCompact = isCompact
-        )
-    }
-}
-
-/**
- * Displays course icon, title and description in a Material Design Card.
- */
-@Composable
-fun CourseItem(
-    modifier: Modifier,
-    isCompact: Boolean,
-    courseWithScore: CourseWithScore,
-    serverUrl: String,
-    authorizationToken: String,
-    onClick: () -> Unit
-) {
-    val currentPoints = courseWithScore.totalScores.studentScores.absoluteScore
-    val maxPoints = courseWithScore.totalScores.maxPoints
-
-    val currentPointsFormatted = remember(currentPoints) {
-        CoursePointsDecimalFormat.format(currentPoints)
-    }
-    val maxPointsFormatted = remember(maxPoints) {
-        CoursePointsDecimalFormat.format(maxPoints)
-    }
-
-    val progress = if (maxPoints == 0f) 0f else currentPoints / maxPoints
-
-    val progressPercentFormatted = remember(progress) {
-        DecimalFormat.getPercentInstance().format(progress)
-    }
-
-    if (isCompact) {
-        CompactCourseItemHeader(
-            modifier = modifier,
-            course = courseWithScore.course,
-            serverUrl = serverUrl,
-            authorizationToken = authorizationToken,
-            onClick = onClick,
-            compactCourseHeaderViewMode = CompactCourseHeaderViewMode.EXERCISE_AND_LECTURE_COUNT,
-            content = {
-                Divider()
-
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 8.dp, horizontal = 16.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    LinearProgressIndicator(
-                        modifier = Modifier.weight(1f),
-                        progress = progress,
-                        trackColor = MaterialTheme.colorScheme.onPrimary
-                    )
-
-                    CourseProgressText(
-                        modifier = Modifier,
-                        currentPointsFormatted = currentPointsFormatted,
-                        maxPointsFormatted = maxPointsFormatted,
-                        progressPercentFormatted = progressPercentFormatted
-                    )
-                }
-            }
-        )
-    } else {
-        ExpandedCourseItemHeader(
-            modifier = modifier,
-            course = courseWithScore.course,
-            serverUrl = serverUrl,
-            authorizationToken = authorizationToken,
-            onClick = onClick,
-            content = {
-                Box(
-                    modifier = Modifier
-                        .weight(1f)
-                        .aspectRatio(1f)
-                        .align(Alignment.CenterHorizontally)
-                ) {
-                    CircularCourseProgress(
-                        modifier = Modifier
-                            .fillMaxSize(0.8f)
-                            .align(Alignment.Center),
-                        progress = progress,
-                        currentPointsFormatted = currentPointsFormatted,
-                        maxPointsFormatted = maxPointsFormatted,
-                        progressPercentFormatted = progressPercentFormatted
-                    )
-                }
-
-                CourseExerciseAndLectureCount(
-                    modifier = Modifier
-                        .align(Alignment.CenterHorizontally)
-                        .padding(vertical = 8.dp),
-                    exerciseCount = courseWithScore.course.exercises.size,
-                    lectureCount = courseWithScore.course.lectures.size,
-                    textStyle = MaterialTheme.typography.bodyLarge.copy(textAlign = TextAlign.Center),
-                    alignment = Alignment.CenterHorizontally
-                )
-            },
-            rightHeaderContent = { }
-        )
-    }
-}
-
-@Composable
-private fun CircularCourseProgress(
-    modifier: Modifier,
-    progress: Float,
-    currentPointsFormatted: String,
-    maxPointsFormatted: String,
-    progressPercentFormatted: String
-) {
-    BoxWithConstraints(modifier = modifier) {
-        val progressBarWidthDp = min(24.dp, maxWidth * 0.1f)
-        val progressBarWidth = with(LocalDensity.current) { progressBarWidthDp.toPx() }
-
-        Canvas(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(progressBarWidthDp)
-        ) {
-            drawArc(
-                color = Color.Green,
-                startAngle = 180f,
-                sweepAngle = 360f * progress,
-                useCenter = false,
-                style = Stroke(width = progressBarWidth)
-            )
-
-            drawArc(
-                color = Color.Red,
-                startAngle = 180f + 360f * progress,
-                sweepAngle = 360f * (1f - progress),
-                useCenter = false,
-                style = Stroke(width = progressBarWidth)
-            )
-        }
-
-        val (percentFontSize, ptsFontSize) = with(LocalDensity.current) {
-            val availableSpace = maxHeight - progressBarWidthDp * 2
-            (availableSpace * 0.2f).toSp() to (availableSpace * 0.1f).toSp()
-        }
-
-        Column(
-            modifier = Modifier.align(Alignment.Center),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = stringResource(
-                    id = R.string.course_overview_course_progress_percentage,
-                    progressPercentFormatted
-                ),
-                style = MaterialTheme.typography.bodyLarge,
-                textAlign = TextAlign.Center,
-                fontSize = percentFontSize,
-                fontWeight = FontWeight.Normal
-            )
-
-            Text(
-                text = stringResource(
-                    id = R.string.course_overview_course_progress_pts,
-                    currentPointsFormatted,
-                    maxPointsFormatted
-                ),
-                style = MaterialTheme.typography.bodyLarge,
-                textAlign = TextAlign.Center,
-                fontSize = ptsFontSize,
-                fontWeight = FontWeight.Bold
-            )
-        }
-    }
-}
-
-@Composable
-private fun CourseProgressText(
-    modifier: Modifier,
-    currentPointsFormatted: String,
-    maxPointsFormatted: String,
-    progressPercentFormatted: String
-) {
-    Text(
-        modifier = modifier,
-        text = stringResource(
-            id = R.string.course_overview_course_progress,
-            currentPointsFormatted,
-            maxPointsFormatted,
-            progressPercentFormatted
-        ),
-        fontSize = 14.sp
-    )
-}
 
 @Composable
 private fun DashboardEmpty(modifier: Modifier, onClickSignup: () -> Unit) {
